+++ templates/_base.html
@@ -115,7 +115,7 @@
             {% if messages %}
             {% for category, message in messages %}
               <div data-alert class="alert alert-{{category}} alert-dismissible show">
-                  {{ message|urlize(30) }}
+                  {{ message|safe }}
                 <button type="button" class="close" data-dismiss="alert" aria-label="Close">
                   <span aria-hidden="true">&times;</span>
                 </button>
